Move omit utils out of emotion-utils
